Use JIT_COMPILE rather than AOT_LOAD resolve flag
The compiler uses J9_RESOLVE_FLAG_AOT_LOAD_TIME for compile time resolve
when doing an AOT load. However, semantically, this should be the same
as J9_RESOLVE_FLAG_JIT_COMPILE_TIME as an AOT compile is, form an
external (to the compiler) point of view, the same as a JIT compile.
Therefore, having semantics in the load of a body than what was used to
generate the body is probably not a good idea. This commit replaces all
uses of the AOT_LOAD resolve flag with the JIT_COMPILE resolve flag.

As an aside, it turns out the AOT_LOAD resolve flag dates back to a time
when the JVM supported simple AOT loads from JXEs; in that mode,
the JIT wasn't necessarily loaded.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>