Object sampling hook object ptr is now a returned value
When the hook is triggered the current thread will have VMAccess but
it may release VMAccess as long as it properly handles that the object
may move while VMAccess has been released. The users of the hook now
have to ensure that the object pointer is properly returned if a GC
may happen.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>