Set prexArg for method handle thunks
The receiver of a method handle thunk is always a known object. Set
prexArg for methodhandle thunks in all cases to propagate the info
without peeking in inlining or global VP in later.

In the special case of a method handle thunk being  mutable callsite
target, the known object info is only available in the inlined body
rather than taken side of the guard.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>