Set TR_silentEnv for AOT tests
Tests such as cmdLineTester_getPid fail because when it reads the output
of the test for some success condition, it instead sees

JIT: env var TR_OptionsAOT is set to forceAOT
JIT: env var TR_Options is set to forceAOT

and fails. This commit sets the TR_silentEnv env var when doing AOT
tests to mitigate this issue.

[ci skip]

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>