Implement version checking for JITaaS
Client sends the version information after it establishes connection with the Server. There will be at the maximum 7 connections in total for any given client.
Server sends back a message which will inform the client whether they are compatible or not. Client will switch to local compilations if incompatible.

[skip ci]
Issue: #4467

Signed-off-by: Harry Yu <harryyu1994@gmail.com>