Port testKitGen to Java to enable it to work on all platforms
The Perl version of testKitGen relies on Perl modules which may not be
available on all platforms we test the JVM on. In particular there are
many problems with Perl on z/OS because of various ASCII/EBCDIC
encoding issues.

To enable testing on all platforms which support Java it naturally
makes sense to implement this tool in Java itself, since we're
guaranteed to always have a boot JDK to build/test with.

This change ports the following files into a single Java file
(TestKitGen.java):

- test/TestConfig/scripts/testKitGen/testKitGen.pl
- test/TestConfig/scripts/testKitGen/makeGenTool/mkgen.pl
- test/TestConfig/scripts/testKitGen/makeGenTool/modesService.pl
- test/TestConfig/scripts/testKitGen/makeGenTool/parseFiles.pl

The port is rather naive as to maintain a near replica of the Perl tool
for easier consumption of the people who work on it. If we make the
transition to making the Java tool the default a lot of
cleanup/improvements can be done. This is outside the scope of the port
however.

As part of this change we've introduced a slight modification to the
existing Perl tool and build system. The following will occur now when
testing:

- TestKitGen.java will be compiled via `$TEST_JDK_HOME`
- TestKitGen Java tool will run with `$TEST_JDK_HOME` using the same
arguments as the Perl counterpart
    - It will output the same files with a ".tkgj" suffix
- Perl tool will run as is
- Perl tool will assert that the checksum of the generated files as
well as the ones with the ".tkgj" suffix are identical
    - This ensures both tools generate exact same content, i.e. the
Java port is a true byte-for-byte port
- Build system will fail if there are any discrepancies and the
offending files will be noted

This change needs to be heavily tested now to ensure no regressions.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>