Skip checks for internal JSR292 methods
`undoCustomizationLogic` is called by MethodHandle thunks, the handle
passed in won't be null.

`FilterArgumentsHandle` thunk and `CollectHandle` thunk use `ILGenMacros` to
emit load from or store into an array. For `FilterArgumentsHandle` the
array is `filters` which is never null. For `CollectHandle` the array is
created right before the store thus it won't be null. The array length
is known to the JIT and the load/store from/into the array is guarantted
to be in the range.

ArrayStore check can also be skipped for `CollectHandle` thunk because the
array is created with a type compatible to the arguments to be
collected.

Also add a change in `TR_J9ByteCodeIlGenerator::genCheckCast()` to allow
checkcast to be skipped.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>