Check name & sig lengths match before memcmp anything
Existing resolvefield code was matching the name (length and
memcmp) before checking the sig (length and memcmp).  This
results in unnecessary memcmps when the names are the same
but the lengths are different.

Update the code to do both length comparisons before doing
any memcmps.

Also remove some functiontable calls as this file is in the VM
and doesn't need to go through the table.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>