Replace the call to Modifier.isVolatile with a boolean field
In `FieldHandle`, we use `Modifier.isVolatile(final_modifiers)` to check
if the field is volatile. `Modifier.isVolatile` is a public method and
has the overhead of HCR guard when inlined. Get rid of the call by
storing the answer in a field. Also remove `final_modifiers` since it's
no longer used after this change.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>