Add testing workaround for Cygwin for TEST_ROOT for Windows
Add a workaround which we will remove before merging this PR. On
Windows, because of the use of Cygwin the `TEST_ROOT` path we generate
in the Java port is different from the one perl generates. For example:

```
ifndef TEST_ROOT
 TEST_ROOT :=
/cygdrive/c/Users/jenkins/workspace/Test_openjdk8_j9_sanity.functional_x86-64_windows_Personal/openjdk-tests/TestConfig/scripts/testKitGen/../../..
endif
```

vs.

```
ifndef TEST_ROOT
 TEST_ROOT :=
C:\Users\jenkins\workspace\Test_openjdk8_j9_sanity.functional_x86-64_windows_Personal\openjdk-tests\TestConfig\scripts\testKitGen\..\..\..
endif
```

For testing purposes the checksum of the generated files will not
match, even though the paths are equivalent. We add a workaround for
Windows to turn non-Cygwin paths into Cygwin paths for the purposes of
testing to ensure we get matching checksums.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>