Modify metronome GC to use object scanners
The metronomes-specific object scanning code was mostly duplicate
logic redundant with the object scanner system used by other
collectors and can be removed. Pointer arraylet scanning is moved
to the MM_RealtimeMarking scheme. The related gc stats are moved
from the JavaMarkStats to OMR in preparation for the migration of
metronome to OMR.

Signed-off-by: Jason Hall <jasonhal@ca.ibm.com>