Emit field load instead of call in expanding MethodHandle invoke
Inliner uses bytecode to find the call node to inline for a call target
not from top level. Having a bytecode shared by multiple calls will
cause the inliner to inline a method into a wrong callsite. This is not
a problem now because MethodHandle invoke is inlined at the top level by
targeted inlining that inlines one level each time. With the upcoming
change to inline the whole MethodHandle chain at once, this will become
a problem.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>