Use a common query to decide what to generate for virtual guards
The virtualGuardHelper() on x86, PPC, and z uses separate but similar
logic to decide whether a NOP virtual guard should be generated.

This change pulls the logic from the OMR project into the common layer
so all platforms can use the same query and make VG NOP decisions in a
consistent way.

This unified query is based on x86 implementation and has HCR guard
added to it.

Signed-off-by: Nigel Yu <yunigel@ca.ibm.com>