(v.0.15.0) Exempt com/sun/proxy classes from validation
When we create a proxy of an interface with MethodHandleProxies and invoke
a default method from the proxied interface's superinterfaces, an access
check in the VM fails in JDK8 because the proxy itself does not have the
superinterface listed as one of its own direct superinterfaces. In JDK11
and later, the generated proxy classes have a flag set that exempts them
from validation. This PR does the same to the proxy classes generated in
JDK8.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>