Fix fetching of unloaded class ranges
[skip ci]
On the client, when vector that will store
class ranges to be sent is initialized, the constructor
used preallocates `unloadedClasses->getNumberOfRanges()`
items and sets them to some default value.
However, when `insert` is called in `unloadedClasses->getRanges`,
it appends class ranges to the end of default values, not replacing
them. Thus, the resulting vector is twice the expected length,
and contains invalid values, which may result in persistent memory
overflow on the server.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>