Initialize fieldClassReg for static unresolved reads
When generating code for a watched field event, if an unresolved
static field is being read, we do not intialize the fieldClassReg
correctly. The fieldClassReg is used to test if a field is being
watched. Failure to initialize this register correctly results in
an invalid test being generated at runtime, causing a bug.

This commit correctly initializes the fieldClassReg for the missing
corner case.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>