Look for right caller method when handling OSR in context of inliner
In many places, OSR-related code in the inliner is calling
getOwningMethodSymbol() on a symbol reference in order to access
information about the calling method.  However, the owning method for
a symbol reference is not necessarily the calling method.

Instead, code should call getInlinedResolvedMethodSymbol() on the
TR::Compilation object for the bytecode's caller index or, if the caller
index is -1, call getMethodSymbol().

These changes are guarded by #if defined(INLINER_OSR_CALLING_METHOD),
so they can be enabled at the same time as corresponding changes in
OMR.  Once that happens, the changes will be made unconditional.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>