Reduce number of messages for AOT compilations
[skip ci]
`ResolvedRelocatableMethod_storeValidationRecordIfNecessary`
is the most frequent message when running in AOT mode.
Reduce the number of them by using class chain cache,
which already exists.
If the defining class is known and the class chain is already cached,
we don't need to make a remote call.
If it's not cached, make a remote call and cache the result.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>