Fix creation of duplicate AOTClassInfos for static fields
When determining whether an AOTClassInfo for a static field
is already present in a local list, we would always compare
ROM class of an existing record with NULL.
This doesn't actually perform the check, and may result in
duplicate AOT class info entries.
The correct thing to do is to compare it with a ROM class of
a defining class of the current field.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>