Fix VSTRS loop indexing error by adding -1 to VLL
This commit fixes an incorrect vector load length used in the vector
string search loop.

The VLL instruction's third operand is a byte-index into vector
registers instead of a load length. The relationship is
Load-index=load-length-1. An incorrect load-index results in
an extra byte being loaded and may lead to crashes if the extra
byte is in an inaccessible area.

Signed-off-by: Nigel Yu <yunigel@ca.ibm.com>