Fix out-of-bounds write in String.getBytes
The previous implementation of String.getBytes(int, int, byte[], int)
was not correctly performing a bounds check on the index parameter, nor
was it performing a null check on the data parameter. It would then
proceed to call String.compress or String.compressedArrayCopy with the
user-provided values, which both assumed their input was trustworthy and
would not perform any null or bounds checks when JITted.

Passing in null for the data parameter and a 32-bit address as the index
parameter would allow an attacker to write arbitrary data to any address
in the first 4GiB of memory. Alternatively, passing in a valid byte
array for the data parameter and an out-of-bounds index for the index
parameter would allow an attacker to write arbitrary data to any address
less than 4GiB past the end of the array. Using this behaviour, it would
be possible for attacker-controlled code to bypass any SecurityManager
and gain full control over the JVM.

Signed-off-by: Ben Thomas <ben@benthomas.ca>