Fix String.indexOf corner case where start == Integer.MAX_VALUE - 1
When calling the `java.lang.String.indexOf` API with a `String`
argument and passing the starting offset to be `Integer.MAX_VALUE - 1`
we encounter an overflow when we test for the following condition:

```
if (start + s2len > s1len) {
    return -1;
}
```

This if statement checks that the there are enough characters from the
starting position the user specified for the substring to occur. i.e.
if the source string is "abcdefg" and the substring is "fghijk" and
start offset is 2 then because the length of the source string is 7 and
the length of the substring is 6, 2 + 6 > 7 so it is impossible for the
substring to be contained within the source string. We can return -1
immediately.

The problem with the if statement is if we pass a starting offset to be
`Integer.MAX_VALUE - 1` which is a valid input the addition in the if
statement overflows and we never enter the if block, even though we
should have had there been no overflow. This causes us to call the
JITHelpers intrinsic which has implicit assumptions about the arguments
and we end up dereferencing memory outside of the heap.

This commit fixes the overflow by using subtraction. We also take this
opportunity to document the implicit assumptions on the JIT intrinsics
that we make use of.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>