Add isClassWorthRemembering filter
This commit adds the means for the SVM to filter out methods that is not
worth remembering. This is a heuristic endeavour. The purpose of this is
to address the issue of validation failures in the load run. More
specifically, this should be used for queries that provide a small
benefit compared to the large cost of causing the load to fail, should
the validation fail.

As such, the filter added in this commit filters out any class that is,
or inherits from, java/lang/Throwable. The reason is because these
classes are only seen on the exception path, which, by definition, is
meant to be exceptional. Therefore, preventing the compiler from
being aware of these classes is unlikely to degrade the mainline code.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>