Implement jcmd
Add front-end class.  Add help functions for both the utility per se and the target commands.
Add test cases.

This is the first part which adds the framework for running diagnostic
commands. It provides the capabailities of jps, jmap, and jstack via the jcmd
utility.  Later changesets will add further commands using this framework and
more options for the front-end command.

[ci skip]

Signed-off-by: Peter Bain <peter_bain@ca.ibm.com>