Add gr11 as dependency before calling VMHelper during Field Watch
Both the generateFillInDataBlockSequenceForUnresolvedField() and generateReportFieldAccessOutlinedInstructions()
functions store their first argument in gr3 and call a VM helper function via a trampoline that uses gr11 to store
the callee address. Before implementing this fix, the contents of gr3 were stored to gr11 before the helper
function call and loaded back to gr3 after. However, since gr11 was being used by the trampoline, after the helper
function call, the wrong data would be loaded back, corrupting gr3 and causing a segmentation fault. In this commit,
gr11 has been added as a dependency to ensure that it used only for the callee address (and not to back up gr3),
and that gr3 is not corrupted.

Signed-off-by: Jackie Midroni <jackie.midroni@mail.utoronto.ca>