Fix ifcmp folding
There can be abitrary bytecodes between an `if` bytecode and its branch
target, some of which don't belong to the if's fall through block. When
folding the `if` to its branch target, we need to turn the it into a
`goto` such that the CFG is correctly constructed. There is no such
problem folding the `if` to its fall through as there is no bytecode in
between.

The following is an example. If we were to fold the `if` at bytecode
index 5 to its branch target 19, current implemenation will result in an
edge between the fall through block for the first `if` and its `else`
block, which means that `s` will always be 2 regardless of the value of
`condition1`. This commit will fix this issue.

```
if (condition1) {
   if (condition2) {
   s = 1;
   }
} else {
  s = 2;
}
return;
```
bytecodes
```
  0: iload_1
  1: ifeq          15
  4: iload_2
  5: ifeq          19
  8: iconst_1
  9: putstatic     #2                  // Field s:I
 12: goto          19
 15: iconst_2
 16: putstatic     #2                  // Field s:I
 19: return
```

closes #6421

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>