Store SVM record only when needed
The TR::ClassInfoIsInitialized SVM record is used to remember whether a
class was initialized during compilation (if the compiler decided to
make this query). The TR_PersistentClassInfo object, which is used to
check for class initialization, is acquired via the
TR_PersistentCHTable::findClassInfoAfterLocking query. When the SVM was
first added to the code base, to simplify things, the
TR::ClassInfoIsInitialized was added to the findClassInfoAfterLocking
method as a preemptive means of remembering the query (since the
majority of the callers of findClassInfoAfterLocking also call
isInitialized soon after).

While this approach was simpler, it had the unfortunate consequence of
increasing the chance of AOT load failures because the compiler now
required classes be initialized even in the cases where, during the
compilation, the compiler did not actually require said classes be
initialized.

This commit changes this behaviour, and makes it so that the compiler
will not preemptively store TR::ClassInfoIsInitialized records.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>