Increment msg sequence number after client cancels compilation
After a compilation request is received, the JITServer will ask
for CHTable updates. At this point the client can cancel the compilation
in progress because classes have been unloaded. The JITServer will
catch an exception, but the sequence number for the stream of
messages that have been processed is not incremented which is
similar to a message being completely lost.
We must increment the sequence number of the messages on the
catch path as well.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>