Track registered J2I thunks on the JITServer
[skip ci]
Previously, whenever a J2I thunk was needed,
JITServer would always create a new one and send it
to the client for registration with the VM.
However, if the thunk has already been registered, we would
just register it again and waste a remote message.

This PR adds tracking of registered thunks on the server per-client.
If a thunk is already registered, do not regenerate it on the server.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>