Replace J9JIT_INTERP_VTABLE_OFFSET with VMEnv::getInterpreterVTableOffset()
J9JIT_INTERP_VTABLE_OFFSET is different on Java8 and Java11. We hit a problem doing
cross platform compilations with a Java11 JITServer and a Java8 JITClient. The Java11 JITServer
should be using Java8 JITClient's J9JIT_INTERP_VTABLE_OFFSET when generating code for it.
The Vtable offset will be sent and cached in the per-client VMInfo struct.

Issue: #6351
[skip ci]

Signed-off-by: Harry Yu <harryyu1994@gmail.com>