Enable jniargtests to run with the JIT
The `-Xint` argument to the JVM was hardcoded in the test playlist
which overrides any preceding `-Xjit` option. The effect of this was
that we are running the jniargtests under the interpreter always. In
addition the way the test is currently structured, running with
`-Xjit:count=0` is not as useful, since many things will be unresolved,
and the native JNI thunks will not have been compiled until the test
has been executed at least once. We correct this by executing the test
block several times.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>