Fix NULL class dereference from VP isAssignableFrom folding
Value propagation contains code to perform compile-time evaluations of a number
of Java APIs. One of these is Class.isAssignableFrom. I have observed a bug
where this API receives an UnresolvedClassConstraint as the constraint for the
second child of the isAssignableFrom call. The code currently assumes that any
ClassType will be able to return a non-NULL pointer from getClass - this is not
the case for an UnresolvedClassConstraint. When a NULL class pointer is
returned the pointer will be dereferenced leading to a crash.

This change restricts folding to the family of ResolvedClassConstraints which
should have a non-NULL class pointer from getClass. Folding will be skipped for
all other kinds of ClassType since these cannot supply a class pointer for
compile-time evaluation.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>