Add a return code for jniargs test
Because of the way the test is structured we are checking for a
non-zero return code to imply there were failure observed. The test is
currently not returning any return code, so even if there are failures
the test will not be marked as "FAILED" via the CI automation.

To fix this we simply return a return code from the summary method to
indicate if there were any failed tests so the Java code can return a
proper exit code indicating a failure.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>