Fix Insertion of allocationFence
When compiling methods that include calls to constructors that initialize private fields in some cases an allocationfence was not inserted after the call.

Previously, isFinalFieldFromSuperClasses was incorrectly comparing the instance size of the superclass and the size of the field offset.
This was due to padding that could occur, causing the size of the superclass instance to be greater than it's actual size within the derived class.

isFinalFieldFromSuperClasses is removed in exchange for newly created jitGetDeclaringClassOfROMField that returns the J9Class which declares the given ROM field.

Signed-off-by: Alen Badel <Alen.Badel@ibm.com>