Deal with FIXMEs for JITServer
[skip ci]
There are multiple JITServer FIXME instances across the codebase.
This commit deals with most of them.

The most significant one is enabling inlining of natives, if the
class of method was loaded by a non-system class loader.
In the past we probably were not able to get the correct client-side
class loader pointer, but now we can do it.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>