Fix issues with connection interruptions
This change fixes improper handling of StreamCancel during out-of-process compilation on the JITServer. Summary of the issue: JITClient sends JITServer a compilationInterrupted message to tell server to abort the compilation. JITClient proceeds to do a new compilation on the same connection. JITServer receives the compilationInterrupted message and proceeds to send JITClient back an errorCode to inform the JITClient that compilation has been interrupted. JITClient sees the errorCode while doing the new compilation and aborts it. The fix is to disable JITServer from sending errorCode back to JITClient when the interruption is caused by the JITClient. Also I have noticed an improper handling of StreamFailure, instead of attempting to send an errorCode back to client when a StreamFailure is triggered during out-of-process compilation, JITServer needs to just clean up the stream and not send anything back to the client on an already dead connection.

This change also includes splitting StreamCancel up into 3 different dedicated exceptions to make the code cleaner. They are StreamClientSessionTerminate, StreamConnectionTerminate and StreamInterrupted. StreamClientSessionTerminate is an exception thrown on the JITServer when JITClient exits. In this case JITServer will clean up the client session data. StreamConnectionTerminate is an exception thrown on the JITServer when JITClient closes the connection. In this case JITServer will clean up the socket. StreamInterrupted is an exception thrown on the JITServer when JITClient interrupts the compilation. In this case, JITServer aborts the out-of-process compilation immediately, re-queue the connection and move on to the next compilation. JITClient will maintain the connection for new compilations in the compilationInterrupted case instead of closing down the connection (what we do previously).

Issue: #6644
[skip ci]

Signed-off-by: Harry Yu <harryyu1994@gmail.com>