Define new Pre- and PostEscapeAnalysis passes
This change defines new optimization passes, preEscapeAnalysis and
postEscapeAnalysis, that are designed to help Escape Analysis (EA) do
a better job under voluntary OSR.

The new preEscapeAnalysis optimization looks for calls to
OSRInductionHelper and adds a fake prepareForOSR call that references
all live auto symrefs and pending pushes.  The goal is that any object
that ends up as a candidate for stack allocation that appears to be
used by such a fake prepareForOSR call can be heapified by EA at that
point.  That will be delivered in a subsequent change to EA.

After EA, the postEscapeAnalysis pass will remove the fake calls to
prepareForOSR.

Submitted on behalf of Andrew Craik <ajcraik@ca.ibm.com>

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>