Improve Escape Analysis under OSR
The preEscapeAnalysis optimization, which was introduced in a previous
change, looks for calls to OSRInductionHelper and adds a fake
prepareForOSR call that references all live auto symrefs and pending
pushes.  Any object that ends up as a candidate for stack allocation
that appears to be used by such a fake prepareForOSR call can be
heapified at that point by Escape Analysis.

During EA itself, those references on prepareForOSR calls are marked
as ignoreable for the purposes of determining whether an object can be
stack allocated.

Submitted on behalf of Andrew Craik <ajcraik@ca.ibm.com>

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>