Modify SCC test to make it compatible with JITServer
`testSCCMLSoftmx_1` from `extended.functional` currently
fails in JITServer mode, because it expects to find SCC hints
produced due to compilation exceeding an expensive memory
compilation threshold. That does not happen for a remote compilation,
because it uses very little memory on the client.
Make the test pass by manually setting the threshold to a small value
that will definitely be exceeded by a remote compilation.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>