Add Java version range for recognized methods expected to be native
For recognized methods that are expected to be native in some versions
of Java, use the xNative or xAnySigNative macros to initialize their
entries in the tables of recognized methods.  These are used to specify
the range of Java versions in which the method is expected to be native.

Also, split table for methods in sun.misc.Unsafe, as many were
implemented using native methods until Java 8, and subsequently
implemented as Java wrappers around native methods.

Similarly, the sun.reflect.Reflection class and
jdk.internal.reflect.Reflection methods that are recognized by the JIT
are native in different versions of Java.  Placed them in separate
tables to handle the differences.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>