Override J9::Compilation::needRelocationsForStatics
[skip ci]
A while ago I added a new query `needRelocationsForStatics`
to OMR, which always returns true.
The plan was to override that query in OpenJ9 so that we don't
require relocation of statics for regular JITServer compilations.
For some reason I did not do that, which led to issue #6688,
i.e. we add a relocation for unresolved static field during a
non-AOT compilation, fail to relocate (because unresolved), and
fail compilation. Fixing this now.
Also, remove a similar query
`J9::CodeGenerator::needRelocationsForStatics`,
since it does the same thing but comp object is better suited for it.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>