Fix clientSessionData nullptr issue
readCompileRequest is not supposed to handle StreamInterrupted.
readCompileRequest is meant to receive messages from buildCompileRequest and buildCompileRequest only. buildCompileRequest does not trigger StreamInterrupted.
writeError is the one that can trigger StreamInterrupted and it's supposed to be handled in read() instead of readCompileRequest.
This issue occurred because of out-of-order messages, where a writeError message was received by readCompileRequest, we should throw a StreamMessageTypeMismatch instead of StreamInterrupted.

Issue: #6687

Signed-off-by: Harry Yu <harryyu1994@gmail.com>