Lower invocation counts for SCC cold runs
Lowering the invocation count results in methods getting compiled
sooner. When the SCC is used, this means that there are more methods
available in the SCC during startup, which improves startup time.
However, this also comes at the cost of increased footprint.

Additionally, it only makes sense to lower the invocation count of the
SCC cold run (ie the populate SCC run), because it just results in more
AOT loads as opposed to having JIT compiled those additional loads;
lowering the invocation count in the SCC warm run (ie the load from SCC
run) risks hurting startup because any methods that weren't compiled in
the first run would now get JIT/AOT compiled sooner; in fact these methods
might not even be all that important as they weren't compiled in the
first run.

This commit extends the existing lower counts for AOT heuristic to apply
by default, unless the default SCC is used - in which case the compiler
falls back respecting the existing heuristic. The reason for this
provision is to not increase footprint for those who are content with
simply using the default SCC.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>