Implement New Lock Reservation scheme on Power
This commit is for a new lock resevation scheme that is currently only
active on power.

The new lock reservation scheme adds a new Learning phase before a lock is
reserved. During the Learning phase a lock reservation cancellation is not
required if a second thread wishes to lock an object. If the Learning
phase is successful, the object can be reserved.

If two threads attempt to lock the same object, then it can never be
reserved again. Under the old system, lock reservation could be attempted
again on the same object even after a reservation on the object was
cancelled.

A system to bypass the new Learning state is added via new initial
lockwords. Data is collected on a per class basis to try and predict
if locked objects of that class will eventually go to Flat or Reserved.
Based on that data, objects might go straight to Flat or Reserved when
they are locked for the first time.

To activate the new lock reservation scheme use the
-XX:+GlobalLockReservation command line option. Without the option, the
lock reservation system used will be very similar to the original system.

Minor changes to x86 codegen were made to accomodate the lockword format
change required to implement the new lock reservation scheme.

This commit also removes much of the unused PPC Monitor Snippet code.

More details can be found in the documentation at:
doc/compiler/ObjectLockword.md

Issue: #8045
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>