Branch to snippet unconditionally for array known to be discontiguous
If the array length in a BNDCHKWithSpineCHK is constant, generated
code was comparing the constant length with the bound, bypassing the
spine check, which is only OK if the array is not discontiguous.

Added a compile-time check of whether the array is discontiguous if the
length is known.  If it is discontiguous, generate an unconditional
branch to the spine check/bound check snippet.

Fixes #6498 and fixes #6773

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>