Restrict fixing simple name to JDK8
In commit 5844a1d4 "Ensure class simple name is a suffix of full name"
getSimpleName was changed to ensure the simple name matched the full name.
This is only relevant in JDK8, as in later versions, the reference
implementation does not ensure the simple name and full name match anymore.

Signed-off-by: Mike Zhang <mike.h.zhang@ibm.com>