Add semi-exhaustive intrinsicIndexOfString* unit tests
The previous set of unit tests for the `intrinsicIndexOfString*` API
were not sufficient to catch a few corner cases when compact strings
are enabled. More specifically the existing tests did not exercise the
`index` parameter of the `java.lang.String.indexOf` APIs and some unit
tests were missed because of the fast paths found in `java.lang.String`.

Rather than relying on `java.lang.String.indexOf` to exercise the
`JitHelpers.intrinsicIndexOfString*` APIs we instead call the APIs
directly and provide a semi-exhaustive set of unit tests for the
various corner cases which we must handle.

Due to the number of inputs the exhaustive set of unit tests would have
totaled into the thousands, and as such an exhaustive set of unit tests
was not feasible to write. Never the less this is an improvement over
the existing tests which are now deprecated in favor of the new set.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>