Fix off-by-one issue in intrinsicIndexOfString* intrinsic on Z
When compact strings are enabled and the input and search strings are
of length 1 the `loadLenReg` will contain a value of 0, signifying we
want to load and compare a single character, because the `VLGV`
instruction is zero-length based (0 means load 1 element). If there is
a match in the first character, the index reported by `VLGV` will also
be 0.

There is a fast-path exit which checks if the index matched is past the
last character we're allowed to read, but it doesn't take into account
that index 0 is valid. As such the following test will fail when
compact strings are enabled:

```
"a".indexOf("a", 0) == 0
```

as it will report -1 instead. We fix this by changing the branch
condition to allow for this corner case.

Fixes: #6671
Fixes: #6778

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>