Port JITServerIProfiler to master
class JITServerIProfiler and class JITClientIProfiler both are derived class of TR_IProfiler that handles the sharing of iprofiler data between JITServer and JITClient. More details on the two classes and their APIs can be found in JITServerIProfiler.hpp/cpp. Part of JITServerCompilationThread.hpp/cpp is included in this change because JITServerIProfiler needs defintions of cacheIProfilerInfo and getCachedIProfilerInfo.

Signed-off-by: Harry Yu <harryyu1994@gmail.com>