Grab TR_J9SharedCache from AOT FE in sampling hook
The JIT FE used in some contexts does not have a TR_J9SharedCache.
In those cases we need to retrieve the AOT FE specifically to ensure
that we can get a TR_J9SharedCache.

When moving the hints APIs from the FE to TR_J9SharedCache[1] most
places in the code that accessed hints via the FE were modified to do so
via TR_J9SharedCache instead, and places where the FE might not
have a TR_J9SharedCache were modified to retrieve the AOT FE first.

The sampling hook is one such place that was previously missed.

[1] dcc7e89b34ecc5d6b94ef81a48f5115e99948b9e

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>