Port JITServerStatisticsThread to master branch
In JITServer the statistics thread is basically the implementation of a
heartbeat mechanism that periodically prints operational statistics to vlog.
The JITServer statistics thread plays the role of the samplingThread in a
normal JVM
It has 3 main duties:
1) Keeps track of time so that other parties can have a cheap way of accessing
   elapsed time
2) Purges the stale client sessions periodically (every 10 seconds)
3) Prints to vlog operational statistics like: number of clients that are
   connected, number of active compilations threads, CPU utilization of the
   JITServer, etc.
The period of the statistics printout is given by _statisticsFrequency.
If this value is 0, no statistics are printed. This value can be changed
with -Xjit:statisticsFrequency=<period-in-ms>
To disable the JITServerStatisticsThread functionality completely
use -Xjit:samplingFrequency=0

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>