Common up the check of TR_OptimizeForSpace , TR_DisableInlineCheckcastand TR_DisableInlineInstanceOf  options
Z and Power used to use the similar structure to check JIToptions 'TR_DisableInlineInstanceOf,' T'R_OptimizeForSpace' and 'TR_DisableInlineCheckcast' in 'instanceof' and 'checkcast’s evaluators before entering the common API 'calculateInstanceOfOrCheckCastSequences' which generate a list of tests for the nodes.

The changes move the check of JIToptions from Z and Power platform to the common API to have it generate 'HelperCall' if the node is 'instanceof' and one of the options 'TR_OptimizeForSpace' and 'TR_DisableInlineInstanceOf' is turned on. The changes also allow the common API to generate 'HelperCall' if the node is 'checkcast/checkCastAndNullCheck' and either 'TR_OptimizeForSpace' or 'TR_DisableInlineCheckcast' is enabled.

And if the node is 'checkCastAndNullCheck', beside the 'HelperCall', the common API would also generate 'NullTest'.

Signed-off-by: simonameng <simonameng97@gmail.com>